+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Custody Service
   - Add support for limiting the number of customers returned via a new header parameter
 ### Changed
+- Change project name from `openwealth-spring-starter` to `openwealth-spring-boot-starter` 
 - Custody Service
   - Improve error handling in `CustomerReader` to use status codes instead of status messages
 ### Fixed